+++ MPowerKit.VirtualizeListView/ItemsLayout.cs
@@ -2,10 +2,15 @@
 
 public interface IItemsLayout
 {
-
+    int PoolSize { get; set; }
 }
 
-public class LinearLayout : IItemsLayout
+public class ItemsLayout : IItemsLayout
+{
+    public int PoolSize { get; set; } = 4;
+}
+
+public class LinearLayout : ItemsLayout
 {
     public double ItemSpacing { get; set; }
 }
